+++ master.py
@@ -127,7 +127,7 @@ class MainWindow(QWidget):
         subitem.setData(0, Qt.UserRole, self.right_panel.count())
         self.right_panel.addWidget(gf.Page(widget=file_conversion.MaskGraphConversion()))
 
-        subitem = QTreeWidgetItem(item, ["File conversion (to mp4 movies)"])
+        subitem = QTreeWidgetItem(item, ["File conversion (lossy preview)"])
         subitem.setData(0, Qt.UserRole, self.right_panel.count())
         self.right_panel.addWidget(gf.Page(widget=file_conversion.ImageMaskConversion()))